[TM-1341] move functions to hook
